Master

+++ src/com/muaavin/webservices/rest/GetPosts_QueryClass.java
@@ -30,93 +30,66 @@ public class GetPosts_QueryClass {
 		MySqlDb db = new MySqlDb();
 		
 		String response = "";
+		Connection conn = db.connect();
+		Statement st = conn.createStatement();
+		ResultSet rs = null;
 		
-		try{
-			 Connection conn = db.connect();
-			
 			 
-			 Statement st = conn.createStatement();
-			 response = "inside try statement :";
-			 ResultSet rs = null;
+		if(Group_name.equals("All"))
+		{
+			//GroupAll_postDetail
+			String sql = "select  * from GroupAll_PostDetailWithFeedBack ;";
+			rs = (ResultSet) st.executeQuery(sql);
+			Post_List = getResultantFacebookPostDetailData( rs, new ArrayList<Post>(), false);
+			rs = (ResultSet) st.executeQuery("select* from GroupAll_TweetDetailWithFeedBack");
+			Post_List = getResultantTwitterPostDetailData(rs,Post_List,true);
+			
+		}
+			 
+		else if(Group_name.equals("A"))
+		{
+			// GroupA_postDetail 
+			String sql = "select  * from GroupA_PostDetailWithFeedBack;";
+			rs = (ResultSet) st.executeQuery(sql);	
+			Post_List = getResultantFacebookPostDetailData( rs, new ArrayList<Post>(), false);
+			rs = (ResultSet) st.executeQuery("select* from GroupA_TweetDetailWithFeedBack");
+			Post_List = getResultantTwitterPostDetailData(rs,Post_List,true);
+		}
+			 
+		else if(Group_name.equals("B"))
+		{	
+			//GroupB_postDetail
+			String sql = "select  * from GroupB_PostDetailWithFeedBack ;";
+			rs = (ResultSet) st.executeQuery(sql);
+			Post_List = getResultantFacebookPostDetailData( rs, new ArrayList<Post>(), false);
+			rs = (ResultSet) st.executeQuery("select* from GroupB_TweetDetailWithFeedBack");
+			Post_List = getResultantTwitterPostDetailData(rs,Post_List,true);
+		}
+			 
+		else if(Group_name.equals("C"))
+		{
+			//GroupC_postDetail 
+			String sql = "select  * from GroupB_PostDetailWithFeedBack ;";
+			rs = (ResultSet) st.executeQuery(sql);
+			Post_List = getResultantFacebookPostDetailData( rs, new ArrayList<Post>(), false);
+			rs = (ResultSet) st.executeQuery("select* from GroupC_TweetDetailWithFeedBack");
+			Post_List = getResultantTwitterPostDetailData(rs,Post_List,true);
+		}
+		System.out.println("QUERY SUCCESSFULL EXECUTED");  conn.close();
 		
-			 if(Group_name.equals("All"))
-			 {
-				String sql = "select  * from GroupAll_postDetail ;";
-				rs = (ResultSet) st.executeQuery(sql);
-			 }
-			 
-			 else if(Group_name.equals("A"))
-			 {
-				String sql = "select  * from GroupA_postDetail ;";
-				rs = (ResultSet) st.executeQuery(sql);			 
-			 }
-			 
-			 else if(Group_name.equals("B")){
-				 String sql = "select  * from GroupB_postDetail ;";
-				 rs = (ResultSet) st.executeQuery(sql);
-			 }
-			 
-			 else if(Group_name.equals("C")){
-				 String sql = "select  * from GroupC_postDetail ;";
-				 rs = (ResultSet) st.executeQuery(sql);
-			 }
-			 
-			 
-			 int i = 1;
-			 
-			 
-			  while(rs.next()) { 
-				  System.out.println("afdg");
-
-				  String post_id = rs.getString("id"); //User_ID
-				  String post_detail = rs.getString("post_Detail");
-				  String Parent_Comment_id  = rs.getString("Parent_Comment_id");
-				  String Comment_ID   = rs.getString("Comment_ID");
-				  String Comment   = rs.getString("Comment");
-				  String Name   = rs.getString("Name");//Name 
-				  String post_image   = rs.getString("Post_Image");
-				  String infringingUser_ProfilePic = rs.getString("Profile_pic");//total_Unlikes
-				  int total_Unlikes =  rs.getInt("total_Unlikes");
-				  
-				  
-				  Post_List.add(new Post(post_id,post_detail,post_image,Parent_Comment_id,Name,infringingUser_ProfilePic,Comment_ID,Comment,total_Unlikes));
-				    
-				  
-				 
-				  
-				  System.out.println("Post_ID  ... "+String.valueOf(post_id));
-				  System.out.println("Post_Detail  ... "+String.valueOf(post_detail));
-				 
-				  i = i + 1;
-				 // break;
-				 }
-			  System.out.println("QUERY SUCCESSFULL EXECUTED");
-			
-			 
-			  conn.close();
-			  
-			}
-			catch(Exception e)
-			{
-				  e.printStackTrace();
-
-				response =  response +" in Eception Method";
-				
-			}
 
 		return AesEncryption.encrypt(Post_List.toString());
-		//return response;
+		
 	}
 	
 	@POST
 	@Path("/DeletePosts")
 	@Produces(MediaType.APPLICATION_JSON)
-	public void deletePosts(@QueryParam("Post_id") String Post_id, @QueryParam("Group_name") String group_name, @QueryParam("User_id") String user_id, @QueryParam("isPostOfSpecificUser") boolean isPostOfSpecificUser) throws Exception
+	public void deletePosts(@QueryParam("Post_id") String Post_id, @QueryParam("Group_name") String group_name, @QueryParam("User_id") String user_id, @QueryParam("isPostOfSpecificUser") boolean isPostOfSpecificUser,@QueryParam("InfringingUserID") String InfringingUserID,@QueryParam("IsTwitterPost") boolean IsTwitterPost) throws Exception
 	{
 		if(Post_id!=null) Post_id = AesEncryption.decrypt(Post_id);
 		if(group_name!=null) group_name = AesEncryption.decrypt(group_name);
 		if(user_id!=null) user_id = AesEncryption.decrypt(user_id);
-		
 		
 		MySqlDb Db = new MySqlDb();
 		
@@ -127,11 +100,12 @@ public class GetPosts_QueryClass {
 			 String sql;
 			 
 			 //if(group_name.equals("All"))
-			 if(getPostCount(Post_id) > 0 )
-			 {
-				 if(isPostOfSpecificUser == true) { deleteUsersPosts( st, Post_id , user_id  );  }
-				 else { deleteAllPosts(st , Post_id  ); }
-			 }
+			 //if(getPostCount(Post_id) > 0 )
+			 //{
+				 if(isPostOfSpecificUser == true) { deleteUsersPosts( st, Post_id , user_id ,IsTwitterPost );  }
+				 else { deleteAllPosts(st , Post_id ,InfringingUserID, IsTwitterPost ); }
+			 //}
+			 
 			 	
 		}
 			 
@@ -165,17 +139,29 @@ public class GetPosts_QueryClass {
 			}
 			
 		  }
-		 catch (SQLException e)
-		 {
+		 catch (SQLException e) { e.printStackTrace(); }
+		 
 		
-			e.printStackTrace();
-		 }
+			  
+		   
 		
 		return  PostCount ;
 	}
 	
-	public void deleteUsersPosts( Statement st, String Post_id , String user_id  ) throws SQLException
+	public void deleteUsersPosts( Statement st, String Post_id , String user_id, boolean  IsTwitterPost  ) throws SQLException
 	{
+		
+		if(IsTwitterPost)
+		{
+			st.executeUpdate("delete from TweetTable where TweetID = '"+Post_id+"' and  User_ID = '"+user_id+"';");
+				
+			System.out.println("Tweet Deleted "+ "Tweet_id " + Post_id + " User id "+user_id);
+			// If Infringing user id does not exist in tweet table
+			// Delete from infringingUser
+			// Delete from TwitterThumbsDown
+			// Delete from TwitterFeedBack
+		}
+		else
 		{
 			 String sql = "";
 			 sql ="delete from infringingUsers where Post_ID =  '"+Post_id+"' and  Profile_Name ='"+user_id+"'; ";
@@ -196,13 +182,30 @@ public class GetPosts_QueryClass {
 			 sql = "delete from Posts_Comments_Table where  Post_ID = '"+Post_id+"';"; 
 			 st.executeUpdate(sql);
 		}
+		System.out.println("TwitterPOst "+IsTwitterPost);
+		
 		
 	}
 	
-	public void deleteAllPosts(Statement st, String Post_id   ) throws SQLException
+	public void deleteAllPosts(Statement st, String Post_id , String InfringingUserID, boolean  IsTwitterPost ) throws SQLException
 	{
+		
+		String sql = "";  boolean RecordAlreadyPresent = false;
+		if(IsTwitterPost)
 		{
-			 String sql = "";
+			 st.executeUpdate("delete from TweetTable where TweetID = '"+Post_id+"';");
+			 st.executeUpdate("delete from TwitterFeedBack where TweetID = '"+Post_id+"';" );
+			 st.executeUpdate("delete from TwitterThumbsDown where TweetID = '"+Post_id+"';");
+			 //st.executeUpdate("delete from TwitterInfringingUsers where User_ID = '"+InfringingUserID+"';");
+			 
+			 ResultSet rs = (ResultSet) st.executeQuery("select * from TweetTable where Infringing_User_ID = '"+InfringingUserID+"';");
+			 
+			 while(rs.next()) {RecordAlreadyPresent = true; break; } 
+			 
+			 if(!RecordAlreadyPresent) { st.executeUpdate( "delete from  TwitterInfringingUsers  where id = '"+InfringingUserID+"';"); } 
+		}
+		else 
+		{	 
 			 sql ="delete from infringingUsers where Post_ID =  '"+Post_id+"' ; ";
 			 st.executeUpdate(sql);
 			 
@@ -220,9 +223,50 @@ public class GetPosts_QueryClass {
 			 
 			 sql = "delete from Posts_Comments_Table where  Post_ID = '"+Post_id+"';"; 
 			 st.executeUpdate(sql);
+		}
 			 
 			 System.out.println("Post Deleted.");
-		}
+			 System.out.println("IsTwitterPost :"+ IsTwitterPost);
+		
+		
+	}
+	
+	public List<Post> getResultantFacebookPostDetailData( ResultSet rs, List<Post> PostDetails, boolean IsTwitterPostDetail) throws SQLException
+	{
+		while(rs.next()) { 
+			  
+			  String post_id = rs.getString("id"); //User_ID
+			  String post_detail = rs.getString("post_Detail");
+			  String Parent_Comment_id  = rs.getString("Parent_Comment_id");
+			  String Comment_ID   = rs.getString("Comment_ID");
+			  String Comment   = rs.getString("Comment");
+			  String Name   = rs.getString("Name");//Name 
+			  String post_image   = rs.getString("Post_Image");
+			  String FeedBackMessage   = rs.getString("Feedback_Message");
+			  String infringingUser_ProfilePic = rs.getString("Profile_pic");//total_Unlikes
+			  int total_Unlikes =  rs.getInt("total_Unlikes");
+			  	
+			  PostDetails.add(new Post(post_id,post_detail,post_image,Parent_Comment_id,Name,infringingUser_ProfilePic,Comment_ID,Comment,total_Unlikes,FeedBackMessage,IsTwitterPostDetail));
+			  //Post_List.add(new Post(post_id,post_detail,post_image,Parent_Comment_id,Name,infringingUser_ProfilePic,Comment_ID,Comment,total_Unlikes,false));
+			 }
+		return PostDetails;
+		
+	}
+	
+	public List<Post> getResultantTwitterPostDetailData( ResultSet rs, List<Post> PostDetails, boolean IsTwitterPostDetail) throws SQLException
+	{
+		while(rs.next()) { 
+			  
+			  String post_id = rs.getString("id"); //User_ID
+			  String post_detail = rs.getString("post_Detail"); 
+			  String post_image   = rs.getString("Post_Image");
+			  String message   = rs.getString("message");
+			  int total_Unlikes =  rs.getInt("total_Unlikes");
+			  	
+			  PostDetails.add(new Post(post_id,post_detail,post_image,"","","","","",total_Unlikes,message,IsTwitterPostDetail));
+			  //Post_List.add(new Post(post_id,post_detail,post_image,Parent_Comment_id,Name,infringingUser_ProfilePic,Comment_ID,Comment,total_Unlikes,false));
+			 }
+		return PostDetails;
 		
 	}
 }
